+++ src/main/java/gift/model/gift/Gift.java
@@ -1,9 +1,7 @@
-package gift.model;
+package gift.model.gift;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
-
-import java.util.List;
 
 @Entity
 @Table(name = "gift")
