+++ src/main/java/gift/model/user/User.java
@@ -1,9 +1,7 @@
-package gift.model;
+package gift.model.user;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
-
-import java.util.List;
 
 @Entity
 @Table(name = "users")
