+++ src/main/java/gift/model/user/UserRequest.java
@@ -1,4 +1,4 @@
-package gift.model;
+package gift.model.user;
 
 public class UserRequest {
