+++ src/main/java/gift/repository/GiftRepository.java
@@ -1,6 +1,6 @@
 package gift.repository;
 
-import gift.model.Gift;
+import gift.model.gift.Gift;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
