+++ src/main/java/gift/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package gift.repository;
 
-import gift.model.User;
+import gift.model.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
