+++ src/main/java/gift/service/UserService.java
@@ -1,13 +1,11 @@
 package gift.service;
 
-import gift.model.User;
-import gift.model.Wish;
+import gift.model.user.User;
 import gift.repository.UserRepository;
 import gift.util.JwtUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.Optional;
 
 @Service
