+++ src/main/java/gift/util/JwtUtil.java
@@ -1,6 +1,6 @@
 package gift.util;
 
-import gift.model.User;
+import gift.model.user.User;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
@@ -15,13 +15,13 @@ public class JwtUtil {
     @Value("${jwt.secretKey}")
     private String secretKey;
 
-    public String getUserEmail(String token)
-    {
+    public String getUserEmail(String token) {
         Claims claims = Jwts.parser().setSigningKey(secretKey).parseClaimsJws(token).getBody();
         String email = claims.getSubject();
         return email;
     }
-    public boolean checkValidateToken(String token){
+
+    public boolean checkValidateToken(String token) {
         try {
             Jwts.parser().setSigningKey(secretKey).parseClaimsJws(token);
             return true;
